Added contex menus to their own helper file
